Select right value from JSON after changes in API
The eCommerce part i Analytics was using the old values, which after
some changes in the API now just returned JSON string - so it is updated
to use Value property. Furthermore Avg. Items Per Purchase is not
neccessary a whole integer, where it could return a lot of decimals -
fixed to 2 decimals. And also made the boxes a bit wider to better fit
larger number like 100K or 1 million.