Localization support for Analytics #4
Added localization support for Analytics. Added translations for da.xml,
en.xml and en_us.xml files. Most labels and texts are translated. And in
Install.cs pick up the translations from the lang files in
/Analytics/App_Plugins/Analytics/config/lang/ folder and add to the lang
files in Umbraco - it already does this for the "analytics" key for the
section name.